// frontend/src/types/movie.ts
export interface IDuration {
  hours: number;
  minutes: number;
  seconds: number;
}

export interface IReview {
  comment?: string;
  createdAt?: Date;
  rating?: number;
  likes?: number;
}

export interface ICrew {
  id?: string;
  name: string;
  role: string;
  profileImage?: string;
}

export interface ICast {
  id?: string;
  name: string;
  as: string;
  profileImage?: string;
}

export interface IMovie {
  _id: string;
  name: string;
  genre: string[];
  trailer?: string;
  rating: number;
  poster: string;
  duration: IDuration;
  description: string;
  language: string;
  releaseDate: Date;
  status: string;
  likes?: number;
  interests?: number;
  is3D?: boolean;
  crew?: ICrew[];
  cast?: ICast[];
  reviews?: IReview[];
  createdAt?: Date;
  updatedAt?: Date;
}